pyproject.toml
[project]
name = "python-scripts"
version = "0.1.0"
description = ""
authors = [
    { name = "Baptiste Lecocq", email = "baptiste.lecocq@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.2,<0.117.0)",
    "openai (>=1.108.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
call-openai = "scripts.openai:main"
